contiene ajustes de respaldo
