+++ src/app/App.js
@@ -1,7 +1,10 @@
-import {encabezado } from "./componentes/Ingreso";
+
+import { encabezado } from "./componentes/cabezera";
+import { ingreso } from "./componentes/Ingreso";
 
 export const app = () => {
   const documento = document.getElementById("root");
 
   documento.appendChild(encabezado());
+  documento.appendChild(ingreso());
 };
