+++ src/app/visual/style.css
@@ -1,12 +1,25 @@
-h1{
- text-align: center;
-}
-body{
-    background-color: aquamarine;
-}
-.principal{
-    background-color: rgb(244, 127, 255);
-}
-h4{
+h1 {
     text-align: center;
-}
+  }
+  body {
+    margin: 0px;
+  }
+  form {
+      margin: auto;
+      display: flex;
+      flex-wrap: wrap;
+      flex-direction: column;
+      align-items: center;
+    }
+    input {
+      width: 200px;
+      height: 25px;
+      margin: 7px;
+      padding: 10px;
+      border-radius: 5px;
+      box-shadow: 2px 2px 8px 4px rgba(0, 0, 0, 0.5);
+      align-items: center;
+    }
+    h4 {
+      text-align: center;
+    }
